Added support for default hotkeys like copy, paste, undo, redo
