+++ app/background.js
@@ -3,8 +3,9 @@
 // It doesn't have any windows which you can see on screen, but we can open
 // window from here.
 
-import { app } from 'electron';
-import devHelper from './helpers/dev';
+import { app, Menu } from 'electron';
+import { devMenuTemplate } from './helpers/dev_menu_template';
+import { editMenuTemplate } from './helpers/edit_menu_template';
 import createWindow from './helpers/window';
 
 // Special module holding environment variables which you declared
@@ -13,7 +14,16 @@ import env from './env';
 
 var mainWindow;
 
+var setApplicationMenu = function () {
+    var menus = [editMenuTemplate];
+    if (env.name !== 'production') {
+        menus.push(devMenuTemplate);
+    }
+    Menu.setApplicationMenu(Menu.buildFromTemplate(menus));
+};
+
 app.on('ready', function () {
+    setApplicationMenu();
 
     var mainWindow = createWindow('main', {
         width: 1000,
@@ -23,7 +33,6 @@ app.on('ready', function () {
     mainWindow.loadURL('file://' + __dirname + '/app.html');
 
     if (env.name !== 'production') {
-        devHelper.setDevMenu();
         mainWindow.openDevTools();
     }
 });
