+++ app/helpers/dev_menu_template.js
@@ -1,0 +1,22 @@
+export var devMenuTemplate = {
+    label: 'Development',
+    submenu: [{
+        label: 'Reload',
+        accelerator: 'CmdOrCtrl+R',
+        click: function () {
+            BrowserWindow.getFocusedWindow().webContents.reloadIgnoringCache();
+        }
+    },{
+        label: 'Toggle DevTools',
+        accelerator: 'Alt+CmdOrCtrl+I',
+        click: function () {
+            BrowserWindow.getFocusedWindow().toggleDevTools();
+        }
+    },{
+        label: 'Quit',
+        accelerator: 'CmdOrCtrl+Q',
+        click: function () {
+            app.quit();
+        }
+    }]
+};
